Add simple gui to start ArDoCo (#2)
This is a first shot of developing a GUI for ArDoCo to resolve issue #1.
Currently, the GUI design is minimalist, simple, and with limited features: The user can select files and add settings like name in the same fashion as with the CLI. The main difference is that the user gets reminded what options they can use and have comfort functionality like file/directory choosers and a Dropdown-menu for supported architecture types.

After pressing "run", ArDoCo is executed (in another Thread). The results are not used in any form within this GUI, yet. Future iterations should present the input and output properly.

Co-authored-by: Dominik Fuchß <develop@fuchss.org>